RIP-399, move big  code-block to dedicated runner class
